requirements.txt
# Core ML and Data Processing
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
datasets>=2.14.0
huggingface-hub>=0.16.0

# Vector Search and Indexing
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=1.7.0

# Data Storage
sqlite3
pyarrow>=12.0.0
sqlalchemy>=2.0.0

# Web Framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# UI Framework
streamlit>=1.25.0
gradio>=3.35.0

# Text Processing and NLP
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0
vaderSentiment>=3.3.2

# HTTP and API clients
httpx>=0.24.0
requests>=2.31.0
openai>=1.0.0
google-generativeai>=0.3.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Docker and deployment
gunicorn>=21.0.0